==> 05_differential_abundance.Rmd <==
# Differential abundance

```{r load_data_differential}
load("resources/data.Rdata")
```

## Create phyloseq object

```{r phyloseq_object, comment="", message=FALSE, warning=FALSE}
# phyloseq object considering structual zeros
phylo_samples <- sample_metadata %>%
  column_to_rownames("sample") %>%
  sample_data() # convert to phyloseq sample_data object

phylo_genome <- genome_counts_filt %>%
  column_to_rownames("genome") %>%
  mutate_all(~ replace(., . == 0, 0.00001)) %>% # add pseudo counts to avoid structural zero issues (note this approach can be improved!) %>%
  mutate_all(~./sum(.)) %>% #apply TSS nornalisation
  otu_table(., taxa_are_rows = TRUE)

phylo_taxonomy <- genome_metadata %>%
  filter(genome %in% rownames(phylo_genome)) %>% # remove structural zeros
  mutate(genome2 = genome) %>% # create a pseudo genome name column
  column_to_rownames("genome2") %>%
  dplyr::select(domain, phylum, class, order, family, genus, species, genome) %>% # add an additional taxonomic level to ensure genome-level analysis (as no all genomes have species-level taxonomic assignments. Otherwise, ANCOMBC2 aggregates analyses per species)
  as.matrix() %>%
  tax_table() # convert to phyloseq tax_table object

phyloseq <- phyloseq(phylo_genome, phylo_taxonomy, phylo_samples)
```

## Run ANCOMBC2

### MAG level

```{r ancom_model_mag, comment="", message=FALSE, warning=FALSE}
differential_abundance <- ancombc2(
  data = phyloseq,
  assay_name = "counts",
  tax_level = NULL, # change to agglomerate analysis to a higher taxonomic range
  fix_formula = "type", # fixed variable(s)
  rand_formula = "(1|individual)",
  p_adj_method = "holm",
  pseudo_sens = TRUE,
  prv_cut = 0,
  lib_cut = 0,
  s0_perc = 0.05,
  group = NULL,
  struc_zero = FALSE,
  neg_lb = FALSE,
  alpha = 0.05,
  n_cl = 2,
  verbose = TRUE,
  global = FALSE,
  pairwise = FALSE,
  dunnet = FALSE,
  trend = FALSE,
  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
  em_control = list(tol = 1e-5, max_iter = 100),
  lme_control = lme4::lmerControl(),
  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100),
  trend_control = NULL
)

# Save differential abundance to data object
save(sample_metadata, 
     genome_metadata, 
     read_counts, 
     genome_counts, 
     genome_counts_filt, 
     genome_tree,
     genome_gifts, 
     phylum_colors,
     data_stats,
     differential_abundance,
     file = "resources/data.Rdata")
```

```{r ancom_stats_mag, comment="", message=FALSE, warning=FALSE}
tax <- data.frame(phyloseq@tax_table) %>%
  rownames_to_column(., "taxon")

differential_abundance_table <- differential_abundance$res %>%
  dplyr::select(genome=taxon, lfc_typefeces, p_typefeces) %>%
  filter(p_typefeces < 0.05) %>%
  dplyr::arrange(p_typefeces) %>%
  merge(., tax, by = "genome")

differential_abundance_table %>%
  select(genome,order,phylum, lfc_typefeces, p_typefeces) %>%
  tt()
```

```{r ancombc_plot_mag, comment="", message=FALSE, warning=FALSE}
ggplot(differential_abundance_table, aes(x = forcats::fct_rev(genome), y = lfc_typefeces, color = phylum)) +
  geom_point(size = 3) +
  scale_color_manual(values = phylum_colors[-c(3,4)]) +
  geom_hline(yintercept = 0) +
  coord_flip() +
  facet_grid(phylum ~ ., scales = "free", space = "free") +
  theme(
    panel.background = element_blank(),
    axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
    axis.title.x = element_blank(),
    strip.background = element_blank(), 
    strip.text = element_blank()
  ) +
  xlab("Genome") +
  ylab("log2FoldChange") +
  guides(col = guide_legend("Phylum"))
```

```{r ancom_rand_volcano, comment="", message=FALSE, warning=FALSE}
differential_abundance$res %>%
  na.omit() %>%
  dplyr::select(genome=taxon, lfc_typefeces, p_typefeces) %>%
  left_join(genome_metadata, by = join_by(genome == genome)) %>%
  mutate(phylum = ifelse(p_typefeces < 0.05, phylum, NA)) %>%
  ggplot(., aes(x = lfc_typefeces, y = -log(p_typefeces), color = phylum)) +
  geom_point() +
  #xlim(c(-10,4)) +
  scale_color_manual(values = phylum_colors[-c(3,4)]) +
  labs(color = "Significance", x = "Log-fold difference between sample types", y = "p-value") +
  theme_classic()
```

### Genus level

```{r ancom_model_genus, comment="", message=FALSE, warning=FALSE}
differential_abundance_genus <- ancombc2(
  data = phyloseq,
  assay_name = "counts",
  tax_level = "genus", # change to agglomerate analysis to a higher taxonomic range
  fix_formula = "type", # fixed variable(s)
  rand_formula = "(1|individual)",
  p_adj_method = "holm",
  pseudo_sens = TRUE,
  prv_cut = 0,
  lib_cut = 0,
  s0_perc = 0.05,
  group = NULL,
  struc_zero = FALSE,
  neg_lb = FALSE,
  alpha = 0.05,
  n_cl = 2,
  verbose = TRUE,
  global = FALSE,
  pairwise = FALSE,
  dunnet = FALSE,
  trend = FALSE,
  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
  em_control = list(tol = 1e-5, max_iter = 100),
  lme_control = lme4::lmerControl(),
  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100),
  trend_control = NULL
)

# Save differential abundance to data object
save(sample_metadata, 
     genome_metadata, 
     read_counts, 
     genome_counts, 
     genome_counts_filt, 
     genome_tree,
     genome_gifts, 
     phylum_colors,
     data_stats,
     differential_abundance,
     differential_abundance_genus,
     file = "resources/data.Rdata")
```

```{r ancom_stats_genus, comment="", message=FALSE, warning=FALSE}
tax <- data.frame(phyloseq@tax_table) %>%
  rownames_to_column(., "taxon")

differential_abundance_genus$res %>%
  filter(nchar(taxon) > 5) %>%
  dplyr::select(taxon=taxon, lfc_typefeces, p_typefeces) %>%
  filter(p_typefeces < 0.05) %>%
  dplyr::arrange(lfc_typefeces) %>%
  left_join(tax %>% select(-c(taxon,genome,species)) %>% unique(), by = join_by(taxon==genus)) %>%
  tt()
```

### Phylum level

```{r ancom_model_phylum, comment="", message=FALSE, warning=FALSE}
differential_abundance_phylum <- ancombc2(
  data = phyloseq,
  assay_name = "counts",
  tax_level = "phylum", # change to agglomerate analysis to a higher taxonomic range
  fix_formula = "type", # fixed variable(s)
  rand_formula = "(1|individual)",
  p_adj_method = "holm",
  pseudo_sens = TRUE,
  prv_cut = 0,
  lib_cut = 0,
  s0_perc = 0.05,
  group = NULL,
  struc_zero = FALSE,
  neg_lb = FALSE,
  alpha = 0.05,
  n_cl = 2,
  verbose = TRUE,
  global = FALSE,
  pairwise = FALSE,
  dunnet = FALSE,
  trend = FALSE,
  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
  em_control = list(tol = 1e-5, max_iter = 100),
  lme_control = lme4::lmerControl(),
  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100),
  trend_control = NULL
)

# Save differential abundance to data object
save(sample_metadata, 
     genome_metadata, 
     read_counts, 
     genome_counts, 
     genome_counts_filt, 
     genome_tree,
     genome_gifts, 
     phylum_colors,
     data_stats,
     differential_abundance,
     differential_abundance_genus,
     differential_abundance_phylum,
     file = "resources/data.Rdata")
```

```{r ancom_stats_phylum, comment="", message=FALSE, warning=FALSE}
differential_abundance_phylum_table <- differential_abundance_phylum$res %>%
  dplyr::select(taxon=taxon, lfc_typefeces, p_typefeces) %>%
  #filter(p_typefeces < 0.05) %>%
  dplyr::arrange(lfc_typefeces)


phylum_colors2 <- read_tsv("https://raw.githubusercontent.com/earthhologenome/EHI_taxonomy_colour/main/ehi_phylum_colors.tsv") %>%
    left_join(differential_abundance_phylum_table, by=join_by(phylum == taxon)) %>%
    arrange(match(phylum, differential_abundance_phylum_table$taxon)) %>%
    select(phylum, colors) %>% 
    unique() %>%
    select(colors) %>%
    pull()

differential_abundance_phylum_table %>%
  filter(p_typefeces < 0.05) %>%
  dplyr::arrange(lfc_typefeces) %>%
  tt()
```

```{r ancombc_plot_phylum, comment="", message=FALSE, warning=FALSE}
differential_abundance_phylum_table %>%
      mutate(taxon=factor(taxon,levels=differential_abundance_phylum_table$taxon)) %>%
      ggplot(aes(x = lfc_typefeces, y = forcats::fct_rev(taxon), fill = taxon)) +
        geom_col(size = 2) +
        scale_fill_manual(values = phylum_colors2) +
        geom_hline(yintercept = 0) +
        geom_vline(xintercept = 4, linetype="dashed", color = "grey", size=1) +
        geom_vline(xintercept = -4, linetype="dashed", color = "grey", size=1) +
        theme(
          panel.background = element_blank(),
          axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
          axis.title.x = element_blank(),
          strip.background = element_blank(), 
          strip.text = element_blank()
        ) +
        labs(x="log2FoldChange", y="Phylum")
```
